src/requirements.txt
flask==2.2.3
flask-restx==1.1.0
pytest==7.4.2
SQLAlchemy==2.0.20
psycopg2==2.9.7
alembic==1.12.0
python-dotenv==1.0.0